rename: QA-Header 파일명을 컴포넌트에 맞게 변경
